requirements.txt
absl-py==2.2.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.5.1
idna==3.10
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==1.26.4
openai-whisper==20250625
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.6
psutil==7.0.0
pvporcupine==3.0.5
PyAudio==0.2.14
pycaw==20240210
pycparser==2.22
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.98
pywin32==310
regex==2024.11.6
requests==2.32.4
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
SpeechRecognition==3.14.3
srt==3.5.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
vosk==0.3.45
websockets==15.0.1
